refactor(test): refactor TaskDbInsert test
- Move ResponseHandler into its own file as TaskDbInsert has too many
responsibilities.

- Split test_run into test on empty response list, and test on non-empty
response list

- Add test_run_insert_fazdb_uptime
